+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/rendering.kt
@@ -23,6 +23,9 @@ internal class DebugRendering(
 ) : Rendering {
 	private var lastRender: TimeMark? = null
 
+	private val staticTextSurfaces = mutableListOf<TextSurface>()
+	private val textSurface = TextSurface(ansiLevel)
+
 	override fun render(node: MosaicNode): CharSequence {
 		var failed = false
 		val output = buildString {
@@ -39,14 +42,13 @@ internal class DebugRendering(
 			appendLine()
 
 			try {
-				val statics = ArrayList<TextSurface>()
-					.also { node.paintStatics(it, ansiLevel) }
-					.map { it.render() }
-				if (statics.isNotEmpty()) {
+				node.paintStatics(staticTextSurfaces, ansiLevel)
+				if (staticTextSurfaces.isNotEmpty()) {
 					appendLine("STATIC:")
-					for (static in statics) {
-						appendLine(static)
+					for (surface in staticTextSurfaces) {
+						appendLine(surface.render())
 					}
+					staticTextSurfaces.clear()
 					appendLine()
 				}
 			} catch (t: Throwable) {
@@ -57,7 +59,10 @@ internal class DebugRendering(
 
 			appendLine("OUTPUT:")
 			try {
-				appendLine(node.paint(ansiLevel).render())
+				textSurface.reset()
+				textSurface.resize(node.width, node.height)
+				node.paint(textSurface)
+				appendLine(textSurface.render())
 			} catch (t: Throwable) {
 				failed = true
 				append(t.stackTraceToString())
@@ -73,9 +78,11 @@ internal class DebugRendering(
 internal class AnsiRendering(
 	private val ansiLevel: AnsiLevel = AnsiLevel.TRUECOLOR,
 ) : Rendering {
-	private val stringBuilder = StringBuilder(100)
-	private val staticSurfaces = ArrayList<TextSurface>()
+	private val stringBuilder = StringBuilder(128)
 	private var lastHeight = 0
+
+	private val staticTextSurfaces = mutableListOf<TextSurface>()
+	private val textSurface = TextSurface(ansiLevel)
 
 	override fun render(node: MosaicNode): CharSequence {
 		return stringBuilder.apply {
@@ -89,6 +96,9 @@ internal class AnsiRendering(
 			}
 
 			fun appendSurface(canvas: TextSurface) {
+				if (canvas.width <= 0 || canvas.height <= 0) {
+					return
+				}
 				for (row in 0 until canvas.height) {
 					canvas.appendRowTo(this, row)
 					if (staleLines-- > 0) {
@@ -101,14 +111,18 @@ internal class AnsiRendering(
 
 			node.measureAndPlace()
 
-			node.paintStatics(staticSurfaces, ansiLevel)
-			for (staticSurface in staticSurfaces) {
-				appendSurface(staticSurface)
+			node.paintStatics(staticTextSurfaces, ansiLevel)
+			if (staticTextSurfaces.isNotEmpty()) {
+				for (surface in staticTextSurfaces) {
+					appendSurface(surface)
+				}
+				staticTextSurfaces.clear()
 			}
-			staticSurfaces.clear()
 
-			val surface = node.paint(ansiLevel)
-			appendSurface(surface)
+			textSurface.reset()
+			textSurface.resize(node.width, node.height)
+			node.paint(textSurface)
+			appendSurface(textSurface)
 
 			// If the new output contains fewer lines than the last output, clear those old lines.
 			for (i in 0 until staleLines) {
@@ -125,7 +139,7 @@ internal class AnsiRendering(
 
 			append(ansiEndSynchronizedUpdate)
 
-			lastHeight = surface.height
+			lastHeight = textSurface.height
 		}
 	}
 }